Basic Multi application basic functionalities are completed
