+++ front_end/src/Authlibrary/AuthRedirect.js
@@ -1,6 +1,5 @@
 import React from "react";
-import {Redirect,useParams,Route,useHistory,Navigate} from 'react-router-dom';
-import jwt_decode from 'jwt-decode';
+import {Redirect,useParams,Route} from 'react-router-dom';
 import { useUser } from "./context/user.context";
 
 export function LoginProtect({Comp,...rest}){
@@ -19,26 +18,14 @@ export function LoginProtect({Comp,...rest}){
 
 }
 
-export function Notfound(){
-    if(localStorage.getItem("user-access")){
-        console.log("Rendering - Not Found");
-      return <Redirect to="/" />
-     }else{
-         return <Redirect to="/login"/>
-     }
-}
 
 export function AuthRedirect(){
     
     let {id_token}=useParams();
-    const {user,setauth}=useUser();
-      console.log(id_token); 
-
-
+    const {_,setauth,Logout}=useUser();
     try{
         localStorage.setItem("user-access",id_token);
         setauth(id_token);
-        
         if(id_token.trim()!==""){
             return <Redirect to="/" />
            }else{
@@ -48,12 +35,6 @@ export function AuthRedirect(){
     }catch(e){
         console.log(e);
     }
-}
-
-
-export function Logout(){
-    localStorage.removeItem("user-access");
-    window.location.href="/login";
 }
 
 export const PrivateRoute=({Comp,...rest})=>{
