+++ front_end/src/Chat/chat.js
@@ -3,8 +3,6 @@ import { Chatview } from "./component/chatview";
 import { SocketProvider } from "./context/socket/socket.context";
 import { Chatsidebar } from "./component/chat.sidebar";
 import { useState } from "react";
-import jwt_decode from 'jwt-decode';
-
 
 export const Chat=()=>{
     const [modal,setmodal]=useState(false);
