+++ front_end/src/index.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import {Chatapp} from './Chat/Chatapp';
 import {Auth} from './User/Registerlogin';
-import {Profile} from './User/Profile';
+import {Profile} from './User/user';
 import { BrowserRouter,Switch,Route} from "react-router-dom";
-import {Notfound,AuthRedirect,Logout,PrivateRoute,LoginProtect} from './Authlibrary/AuthRedirect';
+import {AuthRedirect,PrivateRoute,LoginProtect} from './Authlibrary/AuthRedirect';
 import reportWebVitals from './reportWebVitals';
 import {Toaster} from 'react-hot-toast';
 import { Imageupload } from './Imageuploader/Imageupload';
@@ -14,6 +13,7 @@ import { Catwiki } from './Catwiki/Catwiki';
 import { Shoppingify } from './Shoppingify/shoppingify';
 import { Chat } from './Chat/chat';
 import { UserProvider } from './Authlibrary/context/user.context';
+import { Appoverview } from './Appoverview/Appoverview';
 
 
 ReactDOM.render(
@@ -28,9 +28,9 @@ ReactDOM.render(
 
     <Route exact path="/login/auth/:id_token/:id" component={AuthRedirect} />
 
+    <PrivateRoute exact path="/" Comp={Appoverview}/>
     <PrivateRoute exact path="/profile" Comp={Profile}/>
     <PrivateRoute exact path="/chat" Comp={Chat}/>
-    <PrivateRoute exact path="/app" Comp={Imageupload}/>  
     <PrivateRoute exact path="/imageuploader" Comp={Imageupload}/>
     <PrivateRoute exact path="/imageuploader/:id" component={Imageredirect} />
     <PrivateRoute path="/catwiki" Comp={Catwiki}/>
